zero_shot_theory_generator/requirements.txt
pandas
numpy
scikit-learn
torch
torchvision
transformers
datasets
gradio
shap
lime
python-dotenv
requests
google-generativeai
kaggle